+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Runtime API to access parameters used for energy generation and exchange rate calculations
+
 ### Changed
 
 - Refactor `pallet-energy-fee` to improve structure and maintainability.